requirements.txt
pandas~=2.2.2
matplotlib~=3.9.1
pillow~=10.3.0
kaggle~=1.7.4.5
torch~=2.7.1
torchvision~=0.22.1
numpy~=1.26.4
opencv-python~=4.9.0.80
streamlit~=1.33.0
seaborn~=0.12.2
scikit-learn~=1.5.1